pyproject.toml
[build-system]
requires = ["flit_core >=3.2,<4"]
build-backend = "flit_core.buildapi"

[project]
name = "pip-sbom"
authors = [{name = "Seth Michael Larson", email = "sethmichaellarson@gmail.com"}]
keywords = ["packaging", "pip", "sbom"]
readme = "README.md"
license = {file = "LICENSE"}
classifiers = [
    "Development Status :: 2 - Pre-Alpha",
    "Environment :: Console",
    "Intended Audience :: Developers",
    "Intended Audience :: System Administrators",
    "License :: OSI Approved :: MIT License",
    "Topic :: Security",
]
dynamic = ["version", "description"]
requires-python = ">=3.8"
dependencies = ["packaging", "spdx-tools", "cyclonedx-python-lib", "pip"]

[project.scripts]
pip-sbom = "pip_sbom._internal.cli:cli"
